Fix final test errors to complete Phase 5 merge
- Added missing Flutter services imports for integration tests
- Fixed findsAtLeastOneWidget matcher to findsAtLeastNWidgets(1)
- Added explicit Future<void> type annotations
- Fixed HrvReading field access through metrics object
- Fixed num to double type conversion in mock data service

All errors resolved - build now compiles cleanly.

+++ test/support/mock_data_service.dart
@@ -42,7 +42,7 @@ class MockDataService {
       // Add natural heart rate variability
       final variation = (_random.nextDouble() - 0.5) * 100; // ±50ms variation
       final rrInterval = (baseRr + variation).clamp(400, 1500);
-      rrIntervals.add(rrInterval);
+      rrIntervals.add(rrInterval.toDouble());
     }
 
     // Calculate realistic metrics or use override
